wrap the illustrations into components
