+++ .changeset/eleven-moose-care.md
@@ -1,0 +1,5 @@
+---
+"@mogeko/bookworm-web": patch
+---
+
+Wrap the illustrations into components.
