+++ apps/web/app/layout.tsx
@@ -5,8 +5,6 @@ import type { Metadata } from "next";
 import { siteConfig } from "@/config/site";
 import { sans } from "@/lib/font";
 import { cn } from "@/lib/utils";
-import { SiteFooter } from "@/components/site-footer";
-import { SiteHeader } from "@/components/site-header";
 import { ThemeProvider } from "@/components/theme-wrapper";
 import { Toaster } from "@/components/toaster";
 
@@ -26,11 +24,7 @@ const RootLayout: React.FC<React.PropsWithChildren> = ({ children }) => {
         )}
       >
         <ThemeProvider defaultTheme={{ mode: "system", color: "zinc" }}>
-          <div className="relative flex min-h-screen flex-col">
-            <SiteHeader />
-            {children}
-            <SiteFooter />
-          </div>
+          <div className="relative flex min-h-screen flex-col">{children}</div>
         </ThemeProvider>
         <Toaster />
       </body>
